// src/types/FormikTypes.ts
export interface IErrorState {
  firebase: string;
  email: string;
  password: string;
  passwordConfirmation: string;
  firstname: string;
  lastname: string;
}

export interface FormValues {
  firstname: string;
  lastname: string;
  email: string;
  password: string;
  passwordConfirmation: string;
}